@@ -1,22 +1,33 @@
-import { RouterModule } from '@angular/router'; //import router module
+import { NgModule } from '@angular/core';
+import { RouterModule,PreloadAllModules } from '@angular/router'; //import router module
 
-import { HomeComponent } from './home/home.component';
+//import { HomeComponent } from './home/home.component';
 import { LoginComponent , LoggedInGuard} from './auth';
-export const routing = RouterModule.forRoot([
-  {
-    path: 'home',
-    component: HomeComponent,
-    canActivate: [LoggedInGuard]
-
-  },
+ const routing = RouterModule.forRoot([
   // {
-  //   path: 'login',
-  //   component: LoginComponent
+  //   path: 'home',
+  //   component: HomeComponent,
+  //   canActivate: [LoggedInGuard]
+
   // },
   {
+    path: 'login',
+    component: LoginComponent
+  },
+  {
     path: '',
-    redirectTo: '/home',
+    redirectTo: '/dashboard',
     pathMatch: 'full' //match to this route if the full path is '' emptystring
   }
 
-]);
+], {preloadingStrategy: PreloadAllModules});
+
+@NgModule({
+  imports: [
+    routing
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule {
+
+}
