@@ -6,6 +6,7 @@ const logger = require("morgan"); // use morgan to log requests to the console
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
 const config = require("config");
+const cors = require("cors");
 const appRoutes_1 = require("./routes/appRoutes");
 // ====== utils
 const Logger_1 = require("./utils/Logger");
@@ -30,7 +31,7 @@ class App {
             config: {
                 autoIndex: false // http://mongoosejs.com/docs/guide.html#indexes - prevent auto creation of indexes to prevent performance hit
             } });
-        // this.express.use(cors());
+        this.express.options('*', cors());
         this.express.use(express.static(path.join(__dirname, 'public/dist')));
         this.express.use(logger('dev')); // use morgan to log requests to the console
         this.express.use(bodyParser.json());
