@@ -31,13 +31,14 @@ router.get('/', (req, res) => {
 //redirect client to facebook consent page 
 router.get('/auth-with-facebook', (req, res) => {
     let url = facebook_service_1.FackbookService.getConsentPageUrl();
-    res.redirect(url);
+    res.send(url);
 });
 /*
 1.get code from client and exchange it for an access_token
 2.create/update user with his authentication credentials*/
 router.post('/facebook/code', (req, res) => __awaiter(this, void 0, void 0, function* () {
     let code = req.body.code;
+    console.log('the code :' + code);
     if (code) {
         Logger_1.Logger.d(TAG, 'got code from client >' + code, 'yellow');
         //1.get code from client and exchange it for an access_token:
@@ -73,6 +74,7 @@ router.post('/facebook/code', (req, res) => __awaiter(this, void 0, void 0, func
                     let token = jwt.sign({
                         userId: userDB._id
                     }, 'mySecretForJWTtoken', { expiresIn: '1h' });
+                    Logger_1.Logger.d(TAG, 'sending back the token >' + token, 'yellow');
                     //send 200 with the token
                     res.status(200).json({
                         token: token
@@ -119,7 +121,7 @@ function printFacebookCreds(userCredentials) {
     Logger_1.Logger.d(TAG, 'access_token =' + userCredentials.access_token, 'yellow');
     Logger_1.Logger.d(TAG, 'token_type =' + userCredentials.token_type, 'yellow');
     Logger_1.Logger.d(TAG, 'expires_in =' + userCredentials.expires_in, 'yellow');
-    Logger_1.Logger.d(TAG, '=================== GOT FACEBOOK CREDENTIALS  ===================', 'yellow');
+    Logger_1.Logger.d(TAG, '=================== / GOT FACEBOOK CREDENTIALS  ===================', 'yellow');
 }
 exports.default = router;
 //-------------------------------------SNIPPETS-------------------------
